practical_backend/requirements.txt
amqp==5.0.6
ansible==2.9.6
apache-libcloud==2.8.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
bcrypt==3.1.7
billiard==3.6.4.0
blinker==1.4
Brlapi==0.7.0
celery==5.2.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
command-not-found==0.3
configobj==5.0.6
crcmod==1.7
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.5
djangorestframework==3.13.1
dnspython==1.16.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
isort==5.7.0
Jinja2==2.10.1
jmespath==0.9.4
keyring==18.0.1
kombu==5.2.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
netaddr==0.7.19
netifaces==0.10.4
ntlm-auth==1.1.0
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pep8==1.7.1
pexpect==4.6.0
pgxnclient==1.3
Pillow==7.0.0
prompt-toolkit==3.0.22
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.9.6
pycairo==1.16.2
pycodestyle==2.6.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pykerberos==1.1.14
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
pytz==2019.3
pywinrm==0.3.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selinux==3.0
sil-audit==0.4.4
sil-excel-utils==0.1.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.2.1
wadllib==1.3.3
wcwidth==0.2.5
wrapt==1.12.1
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==3.0.2
xmltodict==0.12.0
zipp==1.0.0
